feat(board): now can sacrify card
Now a player can sacrify a card if he want, but he can do it as much as he wants. Also fix some data sync between players
