+++ Vermines/Assets/Resources/Prefabs/Player/AU_Player.prefab
@@ -1184,7 +1184,7 @@ MonoBehaviour:
   maxCardNumber: 3
   _POV: {fileID: 2528850769732157969}
   _PlayerData: {fileID: 6211597841273035034}
-  _SacrifiedCardList: {fileID: 0}
+  _SacrifiedCardList: {fileID: 233506145529882586}
 --- !u!114 &2528850769732157969
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1430,6 +1430,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   PlayerController: {fileID: 1306004073350071109}
-  PlayedCardList: {fileID: 7587138408122408476, guid: 0388bca40f957734089c8b0a9a575432, type: 3}
+  PlayedCardList: {fileID: 4012910925281518998}
   DiscardedCardList: {fileID: 4278824787023736862}
   SacrifiedCardList: {fileID: 233506145529882586}
