+++ Vermines/Assets/Scripts/Game/Player/Board.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 0388bca40f957734089c8b0a9a575432
-PrefabImporter:
+guid: 45db15aeb1148314eb662a28bb4342e7
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
