+++ Vermines/Assets/Scripts/Game/Player/Board/AList.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class AList : MonoBehaviour
+{
+    // TODO: Implement AList
+}
