+++ Vermines/Assets/Scripts/Game/Player/Board/AList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d328ec821b2377943aecd4d044ea1628
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
