+++ Vermines/Assets/Scripts/Game/Player/Board/DiscardedCardList.cs
@@ -1,13 +1,12 @@
 using Photon.Pun;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using Vermines;
 
-public class PlayedCardList : MonoBehaviourPunCallbacks
+public class DiscardedCardList : MonoBehaviourPunCallbacks
 {
-    public List<PlayedCell> playedCards = new List<PlayedCell>();
-    public float spaceBetween = 1.0f;
-    public int maxCardNumber = 3; // Up to 4 with a specific god
+    public List<CardView> discardedCards = new List<CardView>();
 
     [SerializeField]
     private PhotonView _POV;
@@ -15,13 +14,15 @@ public class PlayedCardList : MonoBehaviourPunCallbacks
     [SerializeField]
     private PlayerData _PlayerData;
 
-    [SerializeField]
-    private SacrifiedCardList _SacrifiedCardList;
+    public void Awake()
+    {
+
+    }
 
     public void Init()
     {
-        Debug.Log("Init PlayedCardList");
-        playedCards.Clear();
+        Debug.Log("Init DiscardedCardList");
+        discardedCards.Clear();
 
         // Enable the view
         gameObject.SetActive(true);
@@ -31,7 +32,7 @@ public class PlayedCardList : MonoBehaviourPunCallbacks
     {
         string syncJson = player.DataToString();
 
-        _POV.RPC("SyncPlayedCard", RpcTarget.OthersBuffered , syncJson);
+        _POV.RPC("SyncPlayedCard", RpcTarget.OthersBuffered, syncJson);
     }
 
     [PunRPC]
@@ -43,21 +44,21 @@ public class PlayedCardList : MonoBehaviourPunCallbacks
 
     private void UpdateReceivedCard(Data data)
     {
-        Debug.Log("Scycing data ... total of player cards -> " + data.PlayedDeck.Cards.Count);
+        Debug.Log("Scycing data ... total of player cards -> " + data.DiscardDeck.Cards.Count);
 
-        if (playedCards.Count < 0 || data == null)
+        if (discardedCards.Count < 0 || data == null)
         {
             return;
         }
 
-        for (int i = 0; i < data.PlayedDeck.Cards.Count; i++)
+        for (int i = 0; i < data.DiscardDeck.Cards.Count; i++)
         {
             CardView card = Instantiate(Resources.Load<GameObject>(Constants.CardPref), transform.position, Quaternion.identity).GetComponent<CardView>();
-            card.SetCard(data.PlayedDeck.Cards[i]);
+            card.SetCard(data.DiscardDeck.Cards[i]);
 
             // Set IsAnonyme to false
             card.GetCard().IsAnonyme = false;
-            card.transform.position = new Vector3(transform.position.x + i * spaceBetween - 6, transform.position.y + 0.5f, transform.position.z);
+            card.transform.position = new Vector3(transform.position.x, (float)(transform.position.y + i * 0.05 + 0.2), transform.position.z);
             card.transform.Rotate(90, 180, 0);
             card.gameObject.SetActive(true);
             card.gameObject.transform.Find("Back").gameObject.SetActive(false);
@@ -67,9 +68,9 @@ public class PlayedCardList : MonoBehaviourPunCallbacks
 
     private void UpdateCardPosition()
     {
-        for (int i = 0; i < playedCards.Count; i++)
+        for (int i = 0; i < discardedCards.Count; i++)
         {
-            playedCards[i].transform.position = new Vector3(transform.position.x + i * spaceBetween - 6 , transform.position.y + 0.5f, transform.position.z);
+            discardedCards[i].transform.position = new Vector3(transform.position.x, (float)(transform.position.y + i * 0.05 + 0.2), transform.position.z);
         }
 
         if (_POV.IsMine)
@@ -84,7 +85,7 @@ public class PlayedCardList : MonoBehaviourPunCallbacks
     {
         if (_POV.IsMine)
         {
-            if (playedCards.Count + 1 > maxCardNumber || card.GetCard() == null)
+            if (card.GetCard() == null)
             {
                 Destroy(card.gameObject);
                 Debug.Log("Played card list is full");
@@ -92,32 +93,18 @@ public class PlayedCardList : MonoBehaviourPunCallbacks
             }
             card.gameObject.SetActive(true);
 
-            _PlayerData.Data.PlayedDeck.AddCard(card.GetCard());
-
-            // TODO: see if store cell instead of card is needed
-            PlayedCell playedCell = card.gameObject.AddComponent<PlayedCell>();
-            playedCell.CardView = card;
-            playedCell.OnClick += SacrifyCard;
-
-
-            playedCards.Add(playedCell);
+            _PlayerData.Data.DiscardDeck.AddCard(card.GetCard());
+            discardedCards.Add(card);
             UpdateCardPosition();
         }
     }
 
-    public void SacrifyCard(PlayedCell card)
-    {
-        Debug.Log("Sacrify card");
-        _SacrifiedCardList.AddCard(card.CardView);
-        RemoveCard(card);
-    }
-
-    public void RemoveCard(PlayedCell card)
+    public void RemoveCard(CardView card)
     {
         if (_POV.IsMine)
         {
-            playedCards.Remove(card);
-            _PlayerData.Data.PlayedDeck.RemoveCard(card.CardView.GetCard());
+            discardedCards.Remove(card);
+            _PlayerData.Data.DiscardDeck.RemoveCard(card.GetCard());
             UpdateCardPosition();
         }
     }
